backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
httpx==0.27.0
pandas==2.2.2
python-dotenv==1.0.1
alembic==1.13.1
aiosqlite==0.22.1
python-multipart==0.0.9